requirements on keys
